move navigation gen to seperate file
