+++ packages/signals/signals/src/core/signal-generators/dom-gen/dom-gen.ts
@@ -79,74 +79,3 @@ export const shouldIgnoreElement = (el: HTMLElement): boolean => {
   }
   return false
 }
-
-function getURLDifferences(url1: URL, url2: URL): ChangedProperties[] {
-  const changed: ChangedProperties[] = []
-  const propertiesToCompare: (keyof URL)[] = [
-    'href',
-    'protocol',
-    'host',
-    'hostname',
-    'port',
-    'pathname',
-    'search',
-    'hash',
-    'username',
-    'password',
-  ]
-
-  for (const property of propertiesToCompare) {
-    if (url1[property] !== url2[property]) {
-      if (property === 'pathname') {
-        changed.push('path')
-      }
-      if (['search', 'hash'].includes(property)) {
-        changed.push(property as ChangedProperties)
-      }
-    }
-  }
-
-  return changed
-}
-
-export class OnNavigationEventGenerator implements SignalGenerator {
-  id = 'navigation'
-
-  register(emitter: SignalEmitter): () => void {
-    // emit navigation signal on page load
-    emitter.emit(
-      createNavigationSignal({
-        action: 'pageLoad',
-        ...this.createCommonFields(),
-      })
-    )
-
-    // emit a navigation signal whenever the URL has changed
-    const urlChange = new URLChangeObservable()
-    urlChange.subscribe(({ previous, current }) =>
-      emitter.emit(
-        createNavigationSignal({
-          action: 'urlChange',
-          prevUrl: previous.href,
-          changedProperties: getURLDifferences(current, previous),
-          ...this.createCommonFields(),
-        })
-      )
-    )
-
-    return () => {
-      urlChange.unsubscribe()
-    }
-  }
-
-  private createCommonFields() {
-    return {
-      // these fields are named after those from the page call, rather than a DOM api.
-      url: location.href,
-      path: location.pathname,
-      hash: location.hash,
-      search: location.search,
-      title: document.title,
-    }
-  }
-}
