+++ packages/signals/signals/src/core/signal-generators/dom-gen/index.ts
@@ -1,14 +1,11 @@
-import {
-  ClickSignalsGenerator,
-  FormSubmitGenerator,
-  OnNavigationEventGenerator,
-} from './dom-gen'
+import { ClickSignalsGenerator, FormSubmitGenerator } from './dom-gen'
 import {
   MutationChangeGenerator,
   OnChangeGenerator,
   ContentEditableChangeGenerator,
 } from './change-gen'
 import { SignalGeneratorClass } from '../types'
+import { OnNavigationEventGenerator } from './navigation-gen'
 
 export const domGenerators: SignalGeneratorClass[] = [
   MutationChangeGenerator,
